// apps/web/src/api-client/model/adminDashboardResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */
import type { AdminDashboardDayRides } from "./adminDashboardDayRides";
import type { AdminDashboardDayRevenue } from "./adminDashboardDayRevenue";
import type { AdminDashboardPopularRoute } from "./adminDashboardPopularRoute";
import type { AdminDashboardUserMetrics } from "./adminDashboardUserMetrics";

export interface AdminDashboardResponse {
    weeklyRides: AdminDashboardDayRides[];
    weeklyRevenue: AdminDashboardDayRevenue[];
    popularRoutes: AdminDashboardPopularRoute[];
    userMetrics: AdminDashboardUserMetrics;
}